命令序列号的类型为int： 修改 IFaceLog.applyCmdSn方法返回类型为int 修改 IFaceLog.isValidCmdSn的参数类型为int 同步修改所有涉及的参数类型

+++ facelog-base/src/codegen/java/net/gdface/facelog/IFaceLogDecorator.java
@@ -252,7 +252,7 @@ public class IFaceLogDecorator implements IFaceLog{
     }
 
     @Override
-    public long applyCmdSn (Token token) {
+    public int applyCmdSn (Token token) {
         return delegate().applyCmdSn(token);
     }
 
@@ -680,7 +680,7 @@ public class IFaceLogDecorator implements IFaceLog{
     }
 
     @Override
-    public boolean isValidCmdSn (long cmdSn) {
+    public boolean isValidCmdSn (int cmdSn) {
         return delegate().isValidCmdSn(cmdSn);
     }
